+++ plugin/web/extensions/sidebar/core/components/content/common/DatasetCard/Field/Fields/point/PointColor.tsx
@@ -4,10 +4,8 @@ import { useCallback, useState } from "react";
 
 import { BaseFieldProps, Cond } from "../types";
 
-import ColorField from "./common/ColorField";
-import ConditionField from "./common/ConditionField";
-import ItemControls from "./common/ItemControls";
-import { ButtonWrapper, Item, Wrapper } from "./common/styled";
+import { ColorField, ConditionField, ItemControls } from "./common";
+import { ButtonWrapper, Item, Wrapper } from "./commonStyles";
 
 const PointColor: React.FC<BaseFieldProps<"pointColor">> = ({ value, editMode, onUpdate }) => {
   const [pointColors, updatePointColors] = useState(value.pointColors);
