+++ plugin/web/extensions/sidebar/core/components/content/common/DatasetCard/Field/Fields/point/common/index.ts
@@ -1,0 +1,9 @@
+import ColorField from "./ColorField";
+import ConditionField from "./ConditionField";
+import Field from "./Field";
+import ItemControls from "./ItemControls";
+import SelectField from "./SelectField";
+import SwitchField from "./SwitchField";
+import TextField from "./TextField";
+
+export { ColorField, ConditionField, Field, ItemControls, SelectField, SwitchField, TextField };
